Bundle core extension on Android
