+++ .changeset/tough-eggs-behave.md
@@ -1,0 +1,7 @@
+---
+'@powersync/react-native': patch
+'@powersync/capacitor': patch
+'@powersync/node': patch
+---
+
+Android: Bundle PowerSync SQLite core extension with package instead of downloading it as a Gradle dependency.
