+++ packages/capacitor/.gitignore
@@ -27,6 +27,7 @@ DerivedData/
 # Built application files
 *.apk
 *.ap_
+android/core
 
 # Files for the ART/Dalvik VM
 *.dex
