+++ packages/capacitor/android/.gitignore
