+++ packages/react-native/android/.gitignore
