+++ packages/react-native/android/build.gradle
@@ -45,11 +45,15 @@ def supportsNamespace() {
 android {
   if (supportsNamespace()) {
     namespace "com.powersync.opsqlite"
+  }
 
-    sourceSets {
-      main {
+  sourceSets {
+    main {
+      if (supportsNamespace()) {
         manifest.srcFile "src/main/AndroidManifestNew.xml"
       }
+
+      jniLibs.srcDirs += ["core/jniLibs"]
     }
   }
 
@@ -88,7 +92,6 @@ repositories {
 def kotlin_version = getExtOrDefault("kotlinVersion")
 
 dependencies {
-  implementation 'com.powersync:powersync-sqlite-core:0.5.3'
   // For < 0.71, this will be from the local maven repo
   // For > 0.71, this will be replaced by `com.facebook.react:react-android:$version` by react gradle plugin
   //noinspection GradleDynamicVersion
